Refactor: Split (clone) createSettings
